feat(StackiPrior): Add implementation
Provide a special stack, which items popping is prioritized for the first n (priorityDeepness
argument) item from the top of the stack, according to the prioritizer function passed as second
argument. This stack also include a popReturned method which is not chainable and return the poped
value. getNextToPop return the "popable" item according to the prioritizer (without popping it from
the stack). The "it" property allow two type of iteration of the stack : one popping each elements
from the internal _stack (deepPop), and the other from a  shallow copy of the internal stack
(shallowPop).